backend/requirements.txt
litellm
google-api-python-client
google-auth
httplib2
google-auth-oauthlib
langchain
langchain-google-genai
langchain-openai
langchain-community
faiss-cpu
tiktoken
pypdf
unstructured
pytesseract
python-dotenv
requests